auth

+++ src/main/java/com/sns/marigold/user/service/InstitutionUserService.java
@@ -2,11 +2,14 @@ package com.sns.marigold.user.service;
 
 import com.sns.marigold.user.dto.InstitutionUserCreateDto;
 import com.sns.marigold.user.dto.InstitutionUserResponseDto;
+import com.sns.marigold.user.dto.InstitutionUserSecurityUpdateDto;
 import com.sns.marigold.user.dto.InstitutionUserUpdateDto;
 import com.sns.marigold.user.entity.InstitutionUser;
 import com.sns.marigold.user.repository.InstitutionUserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
+import java.util.List;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -20,76 +23,79 @@ public class InstitutionUserService {
   private final InstitutionUserRepository institutionUserRepository;
   private final PasswordEncoder passwordEncoder;
 
-  public InstitutionUser findById(Long id) {
-    return institutionUserRepository
-      .findById(id)
-      .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다."));
-  }
-
-  public InstitutionUser findByUsername(String username) {
-    return institutionUserRepository
-      .findByUsername(username)
-      .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다."));
+  public InstitutionUser findById(UUID id) {
+    return institutionUserRepository.findById(id)
+      .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다"));
   }
 
   public InstitutionUser findByEmail(String email) {
     return institutionUserRepository.findByEmail(email)
-      .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다."));
+      .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다"));
   }
 
   // ** create **
 
   @Transactional
-  public InstitutionUserResponseDto create(InstitutionUserCreateDto dto) {
+  public UUID create(InstitutionUserCreateDto dto) {
     String encodedPassword = passwordEncoder.encode(dto.getPassword());
-
     InstitutionUser user = InstitutionUser.builder()
-      .username(dto.getUsername())
       .email(dto.getEmail())
       .password(encodedPassword)
-      .contactPerson(dto.getContactPerson())
-      .contactPhone(dto.getContactPhone())
-      .registrationNumber(dto.getRegistrationNumber())
+      .companyName(dto.getCompanyName())
+      .repName(dto.getRepName())
+      .brn(dto.getBrn())
+      .zipCode(dto.getZipCode())
       .address(dto.getAddress())
+      .detailedAddress(dto.getDetailedAddress())
       .build();
 
     institutionUserRepository.save(user);
-    return InstitutionUserResponseDto.fromUser(user);
+//    return InstitutionUserResponseDto.fromUser(user);
+    return user.getId();
   }
 
   // ** update **
 
   @Transactional
-  public InstitutionUserResponseDto update(Long id, InstitutionUserUpdateDto dto) {
-    int updatedRows = institutionUserRepository.update(id, dto.getUsername(), dto.getEmail(),
-      dto.getContactPerson(),
-      dto.getContactPhone(), dto.getRegistrationNumber(), dto.getAddress());
-
-    if (updatedRows == 0) {
-      throw new EntityNotFoundException("해당 사용자를 찾을 수 없습니다");
-    }
-
+  public InstitutionUserResponseDto update(UUID id, InstitutionUserUpdateDto dto) {
     InstitutionUser user = institutionUserRepository.findById(id)
       .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다."));
 
+    user.update(dto);
+    return InstitutionUserResponseDto.fromUser(user);
+  }
+
+  @Transactional
+  public InstitutionUserResponseDto updateSecurityInfo(UUID id,
+    InstitutionUserSecurityUpdateDto dto) {
+    InstitutionUser user = institutionUserRepository.findById(id)
+      .orElseThrow(() -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다."));
+
+    user.updateSecurityInfo(dto, passwordEncoder);
     return InstitutionUserResponseDto.fromUser(user);
   }
 
   // ** Get **
 
-  public InstitutionUserResponseDto get(Long id) {
-    InstitutionUser user = findById(id);
+  public InstitutionUserResponseDto getById(UUID id) {
+    InstitutionUser user = institutionUserRepository.findById(id).orElseThrow(
+      () -> new EntityNotFoundException("해당 사용자를 찾을 수 없습니다")
+    );
     return InstitutionUserResponseDto.fromUser(user);
   }
 
-  public InstitutionUserResponseDto getByUsername(String username) {
-    InstitutionUser user = findByUsername(username);
-    return InstitutionUserResponseDto.fromUser(user);
+  public List<InstitutionUserResponseDto> getByCompanyName(String companyName) {
+    List<InstitutionUser> users = institutionUserRepository.findByCompanyName(companyName);
+    return users.stream().map(InstitutionUserResponseDto::fromUser).toList();
   }
 
   // ** delete **
 
-  public void delete(Long id) {
+  public void delete(UUID id) {
+    if (!institutionUserRepository.existsById(id)) {
+      throw new EntityNotFoundException("해당 사용자를 찾을 수 없습니다");
+    }
     institutionUserRepository.deleteById(id);
+
   }
 }
